+++ changelog.md
@@ -2,6 +2,8 @@ This release brings the following changes:
 
 * Added encrypted sessions support through an optional password.
 * Updated translations.
+* Updated documentation.
+* Configuration won't be written to disk under certain circunstances.
 
 Important: some anti-virus software may fla parts of this add-on as malicious. Specifically, `url_handler.exe`, which opens `remote://` and `tele://` links. If you don't use this feature, you can safely quarantine or delete the file. Otherwise, you must add it as an exception.